requirements.txt
asgiref==3.2.10
boto3==1.15.1
botocore==1.18.1
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.1.1
django-allauth==0.42.0
django-filter==2.3.0
django-heroku==0.3.1
django-reset-migrations==0.4.0
django-rest-auth==0.9.5
django-s3direct==1.1.8
django-storages==1.10.1
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
gunicorn==20.0.4
idna==2.10
jmespath==0.10.0
oauthlib==3.1.0
psycopg2==2.8.6
PyJWT==1.7.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
whitenoise==5.2.0
